Update slack webhook payload to use blocks
Blocks are the new layout api for slack messages. This updates the
webhook call to remove attachments and add blocks for message
formatting. This also removes some fields we don't use like title,
icon, and color.

All the auto-generated message formats are removed. Messages are
either custom or the commit message is used.